requirements.txt
fastapi
uvicorn[standard]

requests
httpx

ultralytics

pillow
